% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lm.R
\name{AGP_linearmod}
\alias{AGP_linearmod}
\title{Run a linear model with two parameters.}
\usage{
AGP_linearmod(data, dependent_col, independent_col)
}
\arguments{
\item{data}{the dataset you wish to run the linear model from.}

\item{dependent_col}{the response variable - must be numerical. When running the function, it must be put into "quotes."}

\item{independent_col}{the predictor variable - must be categorical. When running the function, it must be put into "quotes."}
}
\value{
return_summary the summary stats for your linear model
}
\description{
You must have one categorical value (with three type possible), and one numerical (continuous) variable.
Returns a summary table of summary statistics.
}
