% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_anova.R
\name{AGP_lm}
\alias{AGP_lm}
\title{Run an ANOVA statistical analysis on two parameters}
\usage{
AGP_lm(df, dependent, independent_cols)
}
\arguments{
\item{df}{the dataset you wish to run the linear model from.}

\item{dependent}{the response variable}

\item{independent_cols}{the predictor variable}
}
\value{
return_summary the summary stats for your linear model
}
\description{
You must have one categorical value (with three type possible), and one numerical (continuous) variable.
Returns a graph of ANOVA analysis and summary table of summary statistics.
}
