% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcshrub.R
\name{calc_den}
\alias{calc_den}
\title{Calculate the density of a shrub}
\usage{
calc_den(len, width, height)
}
\arguments{
\item{len}{Length of the object}

\item{width}{Width of the object}

\item{height}{Height of the object}
}
\value{
density The density of the object (numeric)
}
\description{
Use the length, width, and height of an object to calculate it's density.
Returns density as a floating point number.
}
