% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_mapview.R
\name{mapview}
\alias{mapview}
\title{Create a quick map from lat/long data}
\usage{
mapview(data, lat, long, layer_name)
}
\arguments{
\item{data}{the dataset you desire to create the map from}

\item{lat}{the latitude values in your dataset - this should be its own column. Continuous numerical values.}

\item{long}{the longitude values in your dataset - this should be its own column. Continuous numerical values.}

\item{layer.name}{the title you wish to set to your map.}
}
\value{
samplemap this will be the map that shows up in your viewer screen on the bottom right.
}
\description{
Use your latitude and logitude data for each sample site to make a quick map of your sampling sites.
You will need these packages prior to starting: sf, mapview.
}
