% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_clean.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove NAs from data set}
\usage{
remove_nas(data)
}
\arguments{
\item{data}{the dataset you desire to remove NAs from}
}
\value{
clean the new data set without NAs as a new table
}
\description{
Use your data set as a whole to remove all NAs
}
