% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_boxplot.R
\name{save_plot}
\alias{save_plot}
\title{Create a boxplot of species and their mass, grouped by the species and save as a png file}
\usage{
save_plot(data, x_var, y_var)
}
\arguments{
\item{data}{the dataset you wish to create a boxplot from. Must be in dataframe form}

\item{x_var}{the species column from your dataset (independent variable)}

\item{y_var}{the mass column from your dataset (dependent variable)}
}
\value{
saveplot returning your boxplot
}
\description{
Use species column as x value and mass as y value. Use species column as group value
Returns the entire plot saved to a png file with dimensions of 7x5 and dpi = 300
}
